[effects] Move Traversable[Iterable] instance to effects
